# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Rapidnote API Documentation",
    "version": "1.0.0",
    "description": "A simple API that allows you to have ephemeral chat conversations, scoped by rooms."
  },
  "components": {
    "schemas": {
      "InMessageDTO": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "Message content"
          },
          "username": {
            "type": "string",
            "description": "User that sent the message"
          }
        }
      },
      "OutMessageDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Message ID"
          },
          "content": {
            "type": "string",
            "description": "Message content"
          },
          "username": {
            "type": "string",
            "description": "User that sent the message"
          },
          "timestamp": {
            "type": "string",
            "description": "Message timestamp in ISO format"
          }
        }
      },
      "OutRoomDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Room name"
          },
          "messages": {
            "type": "array",
            "description": "Room messages",
            "items": {
              "$ref": "#/components/schemas/OutMessageDTO"
            }
          },
          "lastActivity": {
            "type": "string",
            "description": "Last room activity in ISO format"
          }
        }
      }
    }
  },
  "paths": {
    "/rooms/{room}": {
      "get": {
        "summary": "Get list of rooms, or messages for a specific room.",
        "description": "Get room list, or messages (and room info) for a specific room by providing the \"room\" path parameter.",
        "parameters": [
          {
            "in": "path",
            "name": "room",
            "description": "The room identifier.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response. Returns room info or messages based on the presence of the \"room\" path parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutRoomDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a message to a room.",
        "description": "Add a message to a specific room. If the room does not exist, it is created. If the room already has the maximum amount of messages (100), the oldest message is removed.",
        "parameters": [
          {
            "in": "path",
            "name": "room",
            "description": "The room identifier. If no room is provided, the default room (root) is used.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Message data to be added to the room.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InMessageDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response. Returns the updated room info.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutRoomDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": []
}